backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
PyJWT==2.8.0
PyMySQL==1.1.0
cryptography==42.0.5
Werkzeug==3.0.2
python-dotenv==1.0.1
